Merge 4.4.0 release to main (#156)
* When polling, raise if an error is returned (#142)

Previously, if an error from the API was returned while polling for updates on a file, it would continue polling. I updated it so that it raises a RequestError if the API returns an error status. If not, it raises a RetryError, which is swallowed by the retry gem's `with_retries` block.

Example of an error that led to continued polling:

```
{"status":"error","error":"File validation error: Uploading of these file types is not allowed.","error_code":"DownloadFileValidationFailedError"}
```

* Sign URL uploads if configured (#139)

Fix namespace for service

Add spec for upload_from_url

Fix requiring of SignatureGenerator and location of tests

* Merge with main and add changelog

* Bump version to 4.4.0

* Apply suggestions from code review

Co-authored-by: Roman Sedykh <rsedykh@uploadcare.com>

---------

Co-authored-by: Josh Schwartzman <josh@dreamseedo.org>
Co-authored-by: Roman Sedykh <rsedykh@uploadcare.com>